chore: rename few methods
