+++ crates/rslint_parser/src/ast/generated.rs
@@ -2,8 +2,3 @@
 pub(super) mod nodes;
 #[rustfmt::skip]
 pub(super) mod tokens;
-
-#[rustfmt::skip]
-pub(super) mod new_nodes;
-#[rustfmt::skip]
-pub(super) mod new_tokens;
