+++ xtask/src/codegen/mod.rs
@@ -1,6 +1,7 @@
 //! Codegen tools for generating Syntax and AST definitions. Derived from Rust analyzer's codegen
-
 mod ast;
+mod ast_src;
+mod generate_nodes;
 mod parser_tests;
 mod syntax;
 
@@ -16,8 +17,8 @@ pub use self::parser_tests::generate_parser_tests;
 // const ERR_INLINE_TESTS_DIR: &str = "crates/ra_syntax/test_data/parser/inline/err";
 
 const SYNTAX_KINDS: &str = "crates/rslint_syntax/src/generated.rs";
-const AST_NODES: &str = "crates/rslint_parser/src/ast/generated/new_nodes.rs";
-const AST_TOKENS: &str = "crates/rslint_parser/src/ast/generated/new_tokens.rs";
+const AST_NODES: &str = "crates/rslint_parser/src/ast/generated/nodes.rs";
+const AST_TOKENS: &str = "crates/rslint_parser/src/ast/generated/tokens.rs";
 
 #[derive(Debug, PartialEq, Eq, Clone, Copy)]
 pub enum Mode {
@@ -44,4 +45,32 @@ fn update(path: &Path, contents: &str, mode: Mode) -> Result<()> {
 	fn normalize(s: &str) -> String {
 		s.replace("\r\n", "\n")
 	}
+}
+
+pub fn to_upper_snake_case(s: &str) -> String {
+	let mut buf = String::with_capacity(s.len());
+	let mut prev = false;
+	for c in s.chars() {
+		if c.is_ascii_uppercase() && prev {
+			buf.push('_')
+		}
+		prev = true;
+
+		buf.push(c.to_ascii_uppercase());
+	}
+	buf
+}
+
+pub fn to_lower_snake_case(s: &str) -> String {
+	let mut buf = String::with_capacity(s.len());
+	let mut prev = false;
+	for c in s.chars() {
+		if c.is_ascii_uppercase() && prev {
+			buf.push('_')
+		}
+		prev = true;
+
+		buf.push(c.to_ascii_lowercase());
+	}
+	buf
 }
