+++ xtask/src/lib.rs
@@ -1,7 +1,6 @@
 //! Codegen tools mostly used to generate ast and syntax definitions. Adapted from rust analyzer's codegen
 
 // pub mod ast;
-pub mod ast_src;
 pub mod codegen;
 pub mod coverage;
 pub mod docgen;
